App: move Color class to its own sources
