+++ src/App/Material.h
@@ -24,163 +24,10 @@
 #ifndef APP_MATERIAL_H
 #define APP_MATERIAL_H
 
-#ifdef __GNUC__
-# include <cstdint>
-#endif
-
-#include <sstream>
-#include <iomanip>
-#include <FCGlobal.h>
+#include <App/Color.h>
 
 namespace App
 {
-
-/** Color class
- */
-class AppExport Color
-{
-public:
-    /**
-     * Defines the color as (R,G,B,A) whereas all values are in the range [0,1].
-     * \a A defines the transparency.
-     */
-    explicit Color(float R=0.0,float G=0.0, float B=0.0, float A=0.0)
-      :r(R),g(G),b(B),a(A){}
-    /**
-     * Does basically the same as the constructor above unless that (R,G,B,A) is
-     * encoded as an unsigned int.
-     */
-    explicit Color(uint32_t rgba)
-    { setPackedValue( rgba ); }
-    /** Copy constructor. */
-    Color(const Color& c)
-      :r(c.r),g(c.g),b(c.b),a(c.a){}
-    /** Returns true if both colors are equal. Therefore all components must be equal. */
-    bool operator==(const Color& c) const
-    {
-        return getPackedValue() == c.getPackedValue();
-        //return (c.r==r && c.g==g && c.b==b && c.a==a);
-    }
-    bool operator!=(const Color& c) const
-    {
-        return !operator==(c);
-    }
-    /**
-     * Defines the color as (R,G,B,A) whereas all values are in the range [0,1].
-     * \a A defines the transparency, 0 means complete opaque and 1 invisible.
-     */
-    void set(float R,float G, float B, float A=0.0)
-    {
-        r=R;g=G;b=B;a=A;
-    }
-    Color& operator=(const Color& c)
-    {
-        r=c.r;g=c.g;b=c.b;a=c.a;
-        return *this;
-    }
-    /**
-     * Sets the color value as a 32 bit combined red/green/blue/alpha value.
-     * Each component is 8 bit wide (i.e. from 0x00 to 0xff), and the red
-     * value should be stored leftmost, like this: 0xRRGGBBAA.
-     *
-     * \sa getPackedValue().
-     */
-    Color& setPackedValue(uint32_t rgba)
-    {
-        this->set((rgba >> 24)/255.0f,
-                 ((rgba >> 16)&0xff)/255.0f,
-                 ((rgba >> 8)&0xff)/255.0f,
-                 (rgba&0xff)/255.0f);
-        return *this;
-    }
-    /**
-     * Returns color as a 32 bit packed unsigned int in the form 0xRRGGBBAA.
-     *
-     *  \sa setPackedValue().
-     */
-    uint32_t getPackedValue() const
-    {
-        return (static_cast<uint32_t>(r*255.0f + 0.5f) << 24 |
-                static_cast<uint32_t>(g*255.0f + 0.5f) << 16 |
-                static_cast<uint32_t>(b*255.0f + 0.5f) << 8  |
-                static_cast<uint32_t>(a*255.0f + 0.5f));
-    }
-    /**
-     * creates FC Color from template type, e.g. Qt QColor
-     */
-    template <typename T>
-    void setValue(const T& q)
-    { set(q.redF(),q.greenF(),q.blueF()); }
-    /**
-     * returns a template type e.g. Qt color equivalent to FC color
-     *
-     */
-    template <typename T>
-    inline T asValue() const {
-        return(T(int(r*255.0f),int(g*255.0f),int(b*255.0f)));
-    }
-    /**
-     * returns color as hex color "#RRGGBB"
-     *
-     */
-    std::string asHexString() const {
-        std::stringstream ss;
-        ss << "#" << std::hex << std::uppercase << std::setfill('0') << std::setw(2) << int(r*255.0f)
-                                                                     << std::setw(2) << int(g*255.0f)
-                                                                     << std::setw(2) << int(b*255.0f);
-        return ss.str();
-    }
-
-    /**
-     * gets color from hex color "#RRGGBB"
-     *
-     */
-    bool fromHexString(const std::string& hex) {
-        if (hex.size() < 7 || hex[0] != '#')
-            return false;
-        // #RRGGBB
-        if (hex.size() == 7) {
-            std::stringstream ss(hex);
-            unsigned int rgb;
-            char c;
-
-            ss >> c >> std::hex >> rgb;
-            int rc = (rgb >> 16) & 0xff;
-            int gc = (rgb >> 8) & 0xff;
-            int bc = rgb & 0xff;
-
-            r = rc / 255.0f;
-            g = gc / 255.0f;
-            b = bc / 255.0f;
-
-            return true;
-        }
-        // #RRGGBBAA
-        if (hex.size() == 9) {
-            std::stringstream ss(hex);
-            unsigned int rgba;
-            char c;
-
-            ss >> c >> std::hex >> rgba;
-            int rc = (rgba >> 24) & 0xff;
-            int gc = (rgba >> 16) & 0xff;
-            int bc = (rgba >> 8) & 0xff;
-            int ac = rgba & 0xff;
-
-            r = rc / 255.0f;
-            g = gc / 255.0f;
-            b = bc / 255.0f;
-            a = ac / 255.0f;
-
-            return true;
-        }
-
-        return false;
-    }
-
-    /// color values, public accessible
-    float r,g,b,a;
-};
 
 /** Material class
  */
